requirements/development.txt
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beanie==1.12.0
beautifulsoup4==4.11.1
black==22.10.0
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cognitojwt==1.4.1
configparser==5.3.0
cpe==1.2.1
cryptography==38.0.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
ecdsa==0.18.0
entrypoints==0.4
exceptiongroup==1.0.0
executing==1.1.1
fastapi==0.85.0
fastapi-utils==0.2.1
fastjsonschema==2.16.2
greenlet==1.1.3
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
iniconfig==1.1.1
ipykernel==6.16.1
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.3
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jwcrypto==1.4.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
motor==3.0.0
multidict==6.0.2
mypy-extensions==0.4.3
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
nested-lookup==0.2.25
notebook==6.5.1
notebook_shim==0.2.0
orjson==3.8.1
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.5.0
pymongo==4.2.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.2.0
python-dateutil==2.8.2
python-decouple==3.6
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
stack-data==0.5.1
starlette==0.20.4
terminado==0.16.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.5.0
typing_extensions==4.3.0
untangle==1.2.1
urllib3==1.26.12
uvicorn==0.18.3
validators==0.20.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
yarl==1.8.1